requirements.txt
# WebSocket Support
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Celery & Task Queue
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# Auth & Security
PyJWT>=2.8.0
cryptography>=41.0.0

# Utilities
requests>=2.31.0
python-decouple>=3.8
geopy>=2.3.0

# Original Requirements
djangorestframework
psycopg2-binary 
django-cors-headers
django_redis
requests
djangorestframework_simplejwt
django-cloudinary-storage 
pillow
cloudinary
pytest
pytest-django
pyjwt
djangorestframework-gis
Django==5.1
gunicorn
django-environ
pytest-factoryboy 
factory_boy
paystackapi
whitenoise
google-auth
django-anymail[amazon-ses]
django-storages[boto3] 
boto3

# new
watchdog>=4.0
# drf-spectacular[sidecar]
drf-spectacular 
drf-spectacular-sidecar

# moitoring
django-prometheus